pyproject.toml
[project]
name = "task-manager-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "mateus-dev-me",email = "mateusmartinsipu2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "tzdata (>=2025.1,<2026.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "psycopg[binary] (>=3.2.5,<4.0.0)"
]

[tool.poetry]
package-mode = false


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
factory-boy = "^3.3.3"
freezegun = "^1.5.1"
testcontainers = "^4.9.1"
ruff = "^0.9.7"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']


[tool.ruff.format]
preview = true
quote-style = 'single'


[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'


[tool.coverage.run]
omit = [
    "**/migrations/*",
    "**/tests/*",
]


[tool.taskipy.tasks]
lint = 'ruff check'
pre_fmt = 'ruff check --fix'
fmt = 'ruff format'
run = 'fastapi dev app/main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=app -vv'
post_test = 'coverage html'
